debian/rules: extract build_dep.tar.gz first
dh_auto_build runs the default target of the Makefile in the top
directory. Extract the build deps before running it, to avoid
downloading the build dependencies again (which fails in OBS as we build
without network there). Remove the redundant "rebar3 compile" and
"rebare3 escriptize" commands, which are part of the default target in
the Makefile.

debian/rules was copied from osmo_dia2gsup, but there it works as-is
because osmo_dia2gsup has no Makefile (-> dh_auto_build is a no-op).

Change-Id: I2bb0b5f608cd5d9a24c2a367425c524447bd002c