diameter: Add spec definitions for S6b (TS 28.273 and dep RFC 5777)
S6b interface HSS<->PGW is defined in TS 28.273 section 9.

TS 28.273 depends on some types defined in RFC 5777, which was missing
so far.

Some enumareated types in RFC 5777 whose enums fields are declared in
yet other dependent RFCs are left with a dummy value, as TODO for the
future if they are ever needed.

Related: OS#6229
Change-Id: I8075dd54a5fec386feecdf797d710a2524413acc