Fix all sets() lookup functions
Previous logic was wrong, since the accumulator function was not really
matching properly as the parameter was shadowed and didn't include the
value from the outer function.
The lookup is fixed here by passing the input lookup key (eg Imsi) as part
of the default/initial value of the fold() function.

Change-Id: I3a4a473b45fa54e857c2092d8d2b1e68ff9535e1